refactor: 快捷 format 放到一起
